A: seedpeer.me
**Fake download buttons.**
To reproduce: search for something, click on a link.

**Links to downloadsafer.com**
There are two. One of the bottom of the website on any page, and on torrent page there is a text: "To start this download, you need a free bitTorrent client like µTorrent." And "µTorrent" is a hyperlink which leads to a 404 page on downloadsafer.com, instead of µTorrent's actual website. downloadsafer.com seems pretty shady to me, partly due to the text under the banner: "Powered by Adaware".