Fixes https://forums.lanik.us/viewtopic.php?f=64&t=44747
Fixes an issue with tag popups not working correctly on the Stack Exchange network of Q&A sites, for tags whose names happen to correspond to those of tracking scripts (e.g. "google-analytics"). Relevant issue report: https://forums.lanik.us/viewtopic.php?f=64&t=44747

Full list of Stack Exchange Q&A domains: https://meta.stackexchange.com/questions/81379/can-we-have-a-list-of-all-the-stack-exchange-domains-somewhere-for-firewall-pur