Fix Anonymous Read Count on Zeroghedge
Hired development firm here from ZeroHedge. The publisher asked us to look into why the anonymous article read counts (tracking popularity only) stopped working on the page and we traced it to this list and GET request. To be clear this get request only retrieves the number of reads for a list of articles by node id. Example:

[Link to GET request that is triggered on front page.](https://www.zerohedge.com/statistics-ajax?entity_ids=668731,669090,669129,669144,669145,669146,669147,669149,669150,669152,669154,669155,669157,669161,669163,669164,669170,669172,669173,669176)

As you can see no private information is exchanged at all. Is this still in violation of your privacy position? A number of readers were complaining about the site being broken and I thought it was worth following up. Thanks, -mf