Update easyprivacy_general.txt
Removed `/timingcg.` because thats the filename we use at RapidSpike for our Real User Monitoring detection. We don't server ads from it and we don't record any PII data. The script is purely for recording timing metrics, browser, device and estimated location of the user. We also use it for identifying data being sent to untrusted hosts in order to protect the website's owners and its users from Magecart attacks.