Hide specific amp elements instead of .amp-unresolved
AMP adds `.amp-unresolved` to every element while its associated JS file is being downloaded. When the JS file is downloaded, the element is "upgraded" and we remove the `.amp-unresolved` class.

I believe easylist is taking advantage of this to hide the elements that have their JS blocked. However, this causes a lot of Cumulative Layout Shift issues, because elements are popping into view when the JS downloads. Eg, all of our `<amp-img>` (which aren't blocked by easylist) start as `display: none` and then once they're upgraded they become visible, causing shifts.

This is a big issue for us, as one of our core goals is for elements to never shift the pages layout. We go through a lot work to make sure every AMP element on the page has the correct size as early as possible.

Worse yet, AMP is working on a new element upgrade system that defers expensive work until it's necessary. We've essentially implemented something just like lazy loading images, but it'll work for every AMP element on the page. This depends on knowing exactly where the element is on the page, so that we can defer the upgrade until it's near viewport. Because easylist is hiding every AMP element on the page, our `IntersectionObserver` will never fire. This means we'll never upgrade the element (and we'll never remove the `.amp-unresolved` class because that happens _after_ upgrade). So every AMP element just stays hidden permanently.

So instead hiding elements indirectly through `.amp-unresolved`, I've gone through and hidden the specific element that's associated with a blocked JS file.

| List                                   | Blocked JS                   | Hidden Elements                              |
|----------------------------------------|------------------------------|----------------------------------------------|
| fanboy_notifications_general_block.txt | - `/amp-user-notification-`  | - `##amp-user-notification` (already hidden) |
:                                        : - `/amp-web-push-`           : - `##amp-web-push-widget` (already hidden)   :
| fanboy_annoyance_general_block.txt     | - `/amp-apester-`            | - `##amp-apester-media` (newly hidden)       |
| fanboy_social_general_block.txt        | - `/amp-social-`             | - `##amp-social-share` (already hidden)      |
| easylist_cookie_general_block.txt      | - `/amp-consent-`            | - `##amp-consent` (already hidden)           |
| easylist_general_block.txt             | - `/amp-ad-`                 | - `##amp-ad` (already hidden)                |
:                                        : - `/amp-auto-ads-`           : - `##amp-custom-ad` (already hidden)         :
:                                        : - `/amp-sticky-ad-` (new)    : - `##amp-sticky-ad` (new)                    :