Add more Admiral hosts
Admiral homepage: https://www.getadmiral.com/.

These hosts reply to a `GET /` with something like:
```
This domain is used by digital publishers to control access to copyrighted
content in accordance with the Digital Millennium Copyright Act and
understand how visitors are accessing their copyrighted content.
```

If the response bodies are hashed with sha256, we get the following:
- `0c04a9bf2ca61527c5b4d1b07ad9f5f5e71984eb6b360eab7d67d8a4a13b7901`
- `2bb5ca9c173d7b5b72c48c6a899014cbc8fc47339ecdc7c40745ffd451a6d07d`
- `53d5d2477cf35502ac35e2c91a1786a7fec9fc408c2dc4dff0c0fecbaf88073d`
- `6b7f9abdc5cbdf259f0e452aa9aeeb3cd2f0edc884ff507479536599b88f841a`
- `a97df4c021b1233952629fbff95c5778c2dd7ecd14eedd71a76eaca7d9d75232`
- `b5344789bcc38f4e88c02440589e8359792c082654849236ba120f05fee24ffe`

The responses only have slight variations - the content is the same.