Restore Angular Material dialog backdrop
Because the selector is too wide, it will break legit behavior of all apps
that use Angular Material and their dialog systems. This include standard
modal dialogs, but also datepickers, auto-completion and so on.

With this selector, it would make it impossible to close the dialog by clicking
outside of it.