A: `https://adops.gr/`
Hi, please add the exception for `https://adops.gr/` on the existing filter `/adops.$domain=`. The page content it’s being blocked by EasyList and the page maintainer said this: "We're an Ad Operations agency based in Greece and our website is for presentation purposes only. There are no ads included. Our site has a problem when the adblock is active, the filter which blocks us is 'easylist' list. Domain with problem is: `https://adops.gr/`"