A:`aniwatch.to`
Hi, there are popups on this domain anywhere you click, and even if the active at that time adserver it's blocked it is being changed really quickly for another one. It is very intrusive and users complained about it several times, and this seems to be a more long term solution.