Allow Snap Inc owned websites to load static assets from snap.com
For unknown to us reason, `snap.com` is blocked in easylist. This is the main domain for "Snap Inc" and hosts many domains, endpoints, services and static files. 

This rule breaks most  of our sites by not loading fonts or allowing us to check if the user should see the cookie consent popup.

As far as we can tell, this rule does not help anyone on the web because we don't serve ads, popups or annoying content from this domain. Let us know why this entry is on easylist, so we can move just that to a different domain. Blocking an entire company's domain feels like overkill.

Here's our previous attempt at unbreaking our sites: https://github.com/easylist/easylist/pull/18537 which didn't get any comments on it.

This PR adds more of Snap Inc's domains to what I presume is the allowlist where this rule doesn't apply.