Isolates this to 3P only
This makes it such that the script is available for loading when
on a first party site such as https://crypto.sardine.ai/buy where
the page breaks if this script doesn't load.

In this case, it's 1P fingerprinting occuring, but in all other
contexts this script gets loaded as a 3P tracker and should be
blocked.